move metrics related to its own file
