+++ source/client.c
@@ -5,8 +5,8 @@
 #include <aws/mqtt/client.h>
 
 #include <aws/mqtt/private/client_impl.h>
-#include <aws/mqtt/private/client_impl_shared.h>
 #include <aws/mqtt/private/mqtt_client_test_helper.h>
+#include <aws/mqtt/private/mqtt_iot_sdk_metrics.h>
 #include <aws/mqtt/private/packets.h>
 #include <aws/mqtt/private/shared.h>
 #include <aws/mqtt/private/topic_tree.h>
